pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-mail>=0.10.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pillow>=11.1.0",
    "sqlalchemy>=2.0.38",
    "werkzeug>=3.1.3",
    "trafilatura>=2.0.0",
    "anthropic>=0.45.2",
]
